small changes (e.g. adding hours worked, moving example user routes calling to new file, etc)
