+++ prisma/schema.prisma
@@ -20,7 +20,7 @@ enum Role {
 
 model User {
   id                 String @id         @default(auto()) @map("_id") @db.ObjectId
-  role               String             @default("")
+  role               Role               @default(VOLUNTEER)
   firstName          String             @default("")
   lastName           String             @default("")
   email              String             @default("")
@@ -40,9 +40,7 @@ model VolunteerDetails {
   hasLicense    Boolean     @default(false)
   speaksEsp     Boolean     @default(false)
   volunteerType String      @default("")
-  daysAvailable String      @default("")
-  timeAvailable String      @default("")
-  basis         String      @default("")
+  hoursWorked   Int         @default(0)
   user          User        @relation(fields: [userId], references: [id])
   userId        String      @unique @db.ObjectId
 }
